refactor(label): make experimental label component an internal one
BREAKING CHANGE: Label is now an internal component
