+++ src/__experimental__/components/number/number.spec.js
@@ -2,7 +2,7 @@ import React from "react";
 import { mount } from "enzyme";
 import Number from "./number.component";
 import Textbox from "../../../components/textbox";
-import Label from "../label";
+import Label from "../../../__internal__/label";
 
 describe("Number Input", () => {
   let wrapper, input, onChangeFn, onKeyDownFn;
