+++ src/__internal__/checkable-input/checkable-input.spec.js
@@ -7,11 +7,9 @@ import { assertStyleMatch } from "../../__spec_helper__/test-utils";
 import CheckableInput from ".";
 import FieldHelpStyle from "../field-help/field-help.style";
 import { FieldLineStyle } from "../../__experimental__/components/form-field/form-field.style";
-import Label from "../../__experimental__/components/label";
+import Label from "../label";
 import HiddenCheckableInputStyle from "./hidden-checkable-input.style";
-import LabelStyle, {
-  StyledLabelContainer,
-} from "../../__experimental__/components/label/label.style";
+import LabelStyle, { StyledLabelContainer } from "../label/label.style";
 import {
   StyledCheckableInput,
   StyledCheckableInputWrapper,
