+++ src/__internal__/label/index.d.ts
