+++ src/__internal__/label/index.js
