+++ src/components/date/date.spec.js
@@ -13,7 +13,7 @@ import Textbox from "../textbox";
 import StyledDateInput from "./date.style";
 import DateHelper from "../../utils/helpers/date/date";
 import { isEdge } from "../../utils/helpers/browser-type-check";
-import Label from "../../__experimental__/components/label";
+import Label from "../../__internal__/label";
 import StyledInputPresentation from "../../__experimental__/components/input/input-presentation.style";
 
 moment.suppressDeprecationWarnings = true;
