+++ src/components/inline-inputs/inline-inputs.spec.js
@@ -4,11 +4,11 @@ import { shallow, mount } from "enzyme";
 import { Row, Column } from "../row";
 import StyledRow from "../row/row.style";
 import StyledColumn from "../row/column/column.style";
-import Label from "../../__experimental__/components/label";
+import Label from "../../__internal__/label";
 import Textbox from "../textbox";
 import InlineInputs from "./inline-inputs.component";
 import { assertStyleMatch } from "../../__spec_helper__/test-utils";
-import { StyledLabelContainer } from "../../__experimental__/components/label/label.style";
+import { StyledLabelContainer } from "../../__internal__/label/label.style";
 
 describe("Inline Inputs", () => {
   let wrapper;
