+++ src/components/select/multi-select/multi-select.spec.js
@@ -13,7 +13,7 @@ import Option from "../option/option.component";
 import SelectList from "../select-list/select-list.component";
 import { StyledSelectList } from "../select-list/select-list.style";
 import Pill from "../../pill";
-import Label from "../../../__experimental__/components/label";
+import Label from "../../../__internal__/label";
 import InputPresentationStyle from "../../../__experimental__/components/input/input-presentation.style";
 
 describe("MultiSelect", () => {
