+++ src/components/switch/switch.spec.js
@@ -10,7 +10,7 @@ import CheckableInput from "../../__internal__/checkable-input";
 import { StyledCheckableInput } from "../../__internal__/checkable-input/checkable-input.style";
 import FieldHelpStyle from "../../__internal__/field-help/field-help.style";
 import HiddenCheckableInputStyle from "../../__internal__/checkable-input/hidden-checkable-input.style";
-import { StyledLabelContainer } from "../../__experimental__/components/label/label.style";
+import { StyledLabelContainer } from "../../__internal__/label/label.style";
 import StyledSwitchSlider from "./__internal__/switch-slider.style";
 import guid from "../../utils/helpers/guid";
 import {
@@ -24,7 +24,7 @@ import { baseTheme } from "../../style/themes";
 import SwitchSliderPanel from "./__internal__/switch-slider-panel.style";
 import SwitchStyle from "./switch.style";
 import SwitchSlider from "./__internal__/switch-slider.component";
-import Label from "../../__experimental__/components/label";
+import Label from "../../__internal__/label";
 
 jest.mock("../../utils/helpers/guid");
 guid.mockImplementation(() => "guid-12345");
